pyproject.toml
[project]
name = "sql-agent-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Indra Prasetya",email = "indraprasetya154@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "langchain-core (>=0.3.43,<0.4.0)",
    "langchain-google-genai (>=2.0.11,<3.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "load-dotenv (>=0.1.0,<0.2.0)",
]

[tool.poetry]
packages = [{include = "sql_agent_ai", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
